fastapi_app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
kafka-python==2.0.2
python-dotenv==1.0.0
starlette==0.27.0
jinja2==3.1.2